fix: update axis mixin
