+++ packages/vchart/src/component/axis/interface/index.ts
@@ -1,0 +1,3 @@
+export * from './common';
+export * from './spec';
+export * from './theme';
