+++ packages/vchart/src/component/axis/mixin/band-axis-mixin.ts
@@ -4,11 +4,15 @@ import type { StringOrNumber } from '../../../typings';
 import type { IEvent } from '../../../event/interface';
 import { ChartEvent } from '../../../constant/event';
 import type { IModel } from '../../../model/interface';
+import type { IAxisLocationCfg } from '../interface';
 
 export interface BandAxisMixin {
   _scale: BandScale;
   _scales: BandScale[];
   _spec: any;
+  _defaultBandPosition: number;
+  _defaultBandInnerPadding: number;
+  _defaultBandOuterPadding: number;
   event: IEvent;
   isSeriesDataEnable: () => boolean;
   computeStatisticsDomain: () => void;
@@ -18,10 +22,21 @@ export interface BandAxisMixin {
 }
 
 export class BandAxisMixin {
+  dataToPosition(values: any[], cfg: IAxisLocationCfg = {}): number {
+    if (values.length === 0 || this._scales.length === 0) {
+      return 0;
+    }
+    const { position, bandScale } = this.getPosition(values);
+    return (
+      position + bandScale.bandwidth() * (cfg.bandPosition ?? this._spec.bandPosition ?? this._defaultBandPosition)
+    );
+  }
+
   valueToPosition(value: any): number {
     const bandStart = this._scale.scale(value);
     return bandStart;
   }
+
   updateGroupScaleRange() {
     let parentScale = this._scale;
     this._scales.forEach((scale, i) => {
@@ -48,6 +63,7 @@ export class BandAxisMixin {
     }
     return { position, bandScale };
   }
+
   calcScales(DEFAULT_BAND_INNER_PADDING: number, DEFAULT_BAND_OUTER_PADDING: number) {
     const { bandPadding, paddingInner, paddingOuter } = this._spec;
     const isBandPaddingArray = isArray(bandPadding);
