+++ packages/vchart/src/component/axis/polar/band-axis.ts
@@ -1,11 +1,10 @@
 import { BandScale } from '@visactor/vscale';
 import { ComponentTypeEnum } from '../../interface/type';
 import { PolarAxis } from './axis';
-import { DEFAULT_BAND_INNER_PADDING, DEFAULT_BAND_OUTER_PADDING, DEFAULT_BAND_POSITION } from './config';
 import { mixin } from '@visactor/vutils';
 import { BandAxisMixin } from '../mixin/band-axis-mixin';
 import type { StringOrNumber } from '../../../typings';
-import type { IAxisLocationCfg, IPolarBandAxisSpec } from './interface';
+import type { IPolarBandAxisSpec } from './interface';
 import { Factory } from '../../../core/factory';
 import { registerAxis } from '../base-axis';
 
@@ -28,16 +27,6 @@ export class PolarBandAxis<T extends IPolarBandAxisSpec = IPolarBandAxisSpec> ex
     return this.computeBandDomain(data);
   }
 
-  dataToPosition(values: any[], cfg: IAxisLocationCfg = {}): number {
-    if (values.length === 0 || this._scales.length === 0) {
-      return 0;
-    }
-    const { position, bandScale } = this.getPosition(values);
-    return (
-      position + bandScale.bandwidth() * (cfg.bandPosition ?? this.getSpec().bandPosition ?? DEFAULT_BAND_POSITION)
-    );
-  }
-
   protected updateScaleRange() {
     super.updateScaleRange();
     this.updateGroupScaleRange();
@@ -55,7 +44,7 @@ export class PolarBandAxis<T extends IPolarBandAxisSpec = IPolarBandAxisSpec> ex
 
   protected initScales() {
     super.initScales();
-    this.calcScales(DEFAULT_BAND_INNER_PADDING, DEFAULT_BAND_OUTER_PADDING);
+    this.calcScales(this._defaultBandInnerPadding, this._defaultBandOuterPadding);
   }
 
   transformScaleDomain() {
