--- packages/vchart/src/component/axis/polar/config.ts
@@ -1,3 +1,0 @@
-export const DEFAULT_BAND_INNER_PADDING = 0;
-export const DEFAULT_BAND_OUTER_PADDING = 0;
-export const DEFAULT_BAND_POSITION = 0;
